client/requirements.txt
typer==0.9.0
pillow==10.2.0
imagehash==4.3.1
numpy==1.26.4
openai==1.13.3
tenacity==8.2.3
python-dotenv==1.0.1
